requirements.txt
discord.py==2.3.2
aiosqlite==0.19.0
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1
